% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{calc_O3_peakseason}
\alias{calc_O3_peakseason}
\title{Calculate O3 peak-season concentration per year and site, based on data as hourly means in rOstluft::format_rolf()}
\usage{
calc_O3_peakseason(data, min_coverage = 9/12)
}
\arguments{
\item{data}{...}

\item{min_coverage}{...}
}
\description{
Calculate O3 peak-season concentration per year and site, based on data as hourly means in rOstluft::format_rolf()
}
\keyword{internal}
