% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{consecutive_months}
\alias{consecutive_months}
\title{Identify relevant months per year for metric calculation based on O3 monthly mean data (used in O3 peak-season calculation)}
\usage{
consecutive_months(starttime, o3_m1)
}
\arguments{
\item{starttime}{...}

\item{o3_m1}{...}
}
\description{
Identify relevant months per year for metric calculation based on O3 monthly mean data (used in O3 peak-season calculation)
}
\keyword{internal}
