% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{convert_interval2}
\alias{convert_interval2}
\title{Copy from rOstluft::convert_interval()}
\usage{
convert_interval2(interval)
}
\arguments{
\item{interval}{...}
}
\description{
Copy from rOstluft::convert_interval()
}
\keyword{internal}
