% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_monitoring_aq_y1}
\alias{data_monitoring_aq_y1}
\title{Air quality monitoring data for typical pollutants within Ostluft}
\format{
## `data_monitoring_aq_y1`
A tibble:
...
}
\source{
<https://www.ostluft.ch/>, <https://www.bafu.admin.ch/bafu/de/home/themen/luft/zustand/daten/nationales-beobachtungsnetz-fuer-luftfremdstoffe--nabel-.html>
}
\usage{
data_monitoring_aq_y1
}
\description{
Air quality monitoring data as yearly values at sites in Ostluft, Switzerland,
for selected pollutants (NO2, O3, PM10, PM2.5, eBC)
from cantonal (Ostluft) and national (NABEL, BAFU & Empa) air quality monitoring networks.
}
\keyword{datasets}
