% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_monitoring_met_d1}
\alias{data_monitoring_met_d1}
\title{Meteorological SMA-monitoring data from MeteoSchweiz}
\format{
## `data_monitoring_met_d1`
A tibble:
...
}
\source{
<https://www.meteoschweiz.ch/>
}
\usage{
data_monitoring_met_d1
}
\description{
Meteorological monitoring data as daily values at selected sites in the Canton of Zürich, Switzerland,
for selected parameters
from MeteoSwiss monitoring networks.
}
\keyword{datasets}
