% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_monitoring_ndep}
\alias{data_monitoring_ndep}
\title{Air quality monitoring period-data for deriving reactive nitrogen deposition to sensitive ecosystems in Ostluft}
\format{
## `data_monitoring_ndep`
A tibble:
...
}
\source{
<https://www.ostluft.ch/>, <https://www.bafu.admin.ch/bafu/de/home/themen/luft/zustand/daten/nationales-beobachtungsnetz-fuer-luftfremdstoffe--nabel-.html>
}
\usage{
data_monitoring_ndep
}
\description{
Concentration or deposition data, respectively, for NH3, NO2, HNO3, aerosol NH4+, aerosol NO3-, bulk deposition of NH4+, bulk deposition of NO3-
at sites in the area of Ostluft, Switzerland,
based on sampling periods at ecosystem reactive nitrogen deposition sites
from the intercantonal (Ostluft) and national (NABEL) air quality monitoring networks.
}
\keyword{datasets}
