% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_monitoring_ndep_y1}
\alias{data_monitoring_ndep_y1}
\title{Compiled data for reactive nitrogen deposition to sensitive ecosystems in Ostluft, based on monitoring data and monitoring-based statistical models}
\format{
## `data_monitoring_ndep_y1`
A tibble:
...
}
\source{
<https://www.ostluft.ch/>, <https://www.bafu.admin.ch/bafu/de/home/themen/luft/zustand/daten/nationales-beobachtungsnetz-fuer-luftfremdstoffe--nabel-.html>, <https://www.wsl.ch/>
}
\usage{
data_monitoring_ndep_y1
}
\description{
Yearly deposition data for various deposition vectors and compounds: NH3, NO2, HNO3, aerosol NH4+, aerosol NO3-, bulk deposition of NH4+, bulk deposition of NO3-
at sites in the area of Ostluft, Switzerland,
based on sampling at ecosystem reactive nitrogen deposition and ammonia monitoring sites at sensitive ecosystems;
from the intercantonal (Ostluft) and national (NABEL) air quality monitoring networks as well as from Swiss federal research institute WSL, in one case.
}
\keyword{datasets}
