% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{max_mean_h8gl}
\alias{max_mean_h8gl}
\title{Calculate daily maximum 8h running-mean O3 concentration based on O3 1h data in rOstluft::format_rolf() (used in O3 peak-season calculation)}
\usage{
max_mean_h8gl(data)
}
\arguments{
\item{data}{...}
}
\description{
Calculate daily maximum 8h running-mean O3 concentration based on O3 1h data in rOstluft::format_rolf() (used in O3 peak-season calculation)
}
\keyword{internal}
