% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{pad2}
\alias{pad2}
\title{Copy from rOstluft::pad() => because this is the only function we need from this package}
\usage{
pad2(data, start_date = NULL, end_date = NULL, drop_last = FALSE)
}
\arguments{
\item{data}{...}

\item{start_date}{...}

\item{end_date}{...}

\item{drop_last}{...}
}
\description{
Copy from rOstluft::pad() => because this is the only function we need from this package
}
\keyword{internal}
