% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{pad_serie2}
\alias{pad_serie2}
\title{Copy from rOstluft::pad_serie()}
\usage{
pad_serie2(serie, start_date = NULL, end_date = NULL, drop_last = FALSE)
}
\arguments{
\item{serie}{...}

\item{start_date}{...}

\item{end_date}{...}

\item{drop_last}{...}
}
\description{
Copy from rOstluft::pad_serie()
}
\keyword{internal}
