% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{prep_site_meta_nabel}
\alias{prep_site_meta_nabel}
\title{Restructure NABEL site metadata}
\usage{
prep_site_meta_nabel(meta)
}
\arguments{
\item{meta}{...}
}
\description{
Restructure NABEL site metadata
}
\keyword{internal}
