% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{prep_site_meta_ostluft}
\alias{prep_site_meta_ostluft}
\title{Restructure Ostluft site metadata}
\usage{
prep_site_meta_ostluft(meta)
}
\arguments{
\item{meta}{...}
}
\description{
Restructure Ostluft site metadata
}
\keyword{internal}
