% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_monitoring_meta}
\alias{prepare_monitoring_meta}
\title{Prepare and merge monitoring site metadata from Ostluft and NABEL for use in scripts}
\usage{
prepare_monitoring_meta(meta_ostluft, meta_nabel)
}
\arguments{
\item{meta_ostluft}{...}

\item{meta_nabel}{...}
}
\description{
Prepare and merge monitoring site metadata from Ostluft and NABEL for use in scripts
}
\keyword{internal}
