% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_monitoring_nabel_y1}
\alias{prepare_monitoring_nabel_y1}
\title{Prepare usable tibble with air pollutant year-statistics exported and read from https://www.arias.ch/ibonline/ib_online.php and restructure the data similar to a standard long-format (see rOstluft::format_rolf())}
\usage{
prepare_monitoring_nabel_y1(
  data,
  metrics = list(Jahresmittel = c("NO2", "PM10", "PM2.5", "EC / Russ"),
    `höchster 98\%-Wert eines Monats` = "O3"),
  keep_incomplete = FALSE,
  tz = "Etc/GMT-1"
)
}
\arguments{
\item{data}{...}

\item{keep_incomplete}{...}

\item{tz}{...}
}
\description{
Prepare usable tibble with air pollutant year-statistics exported and read from https://www.arias.ch/ibonline/ib_online.php and restructure the data similar to a standard long-format (see rOstluft::format_rolf())
}
\keyword{internal}
