% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_monitoring_ostluft_h1}
\alias{prepare_monitoring_ostluft_h1}
\title{Calculate O3 peak-season metric from Ostluft monitoring h1 data}
\usage{
prepare_monitoring_ostluft_h1(data, keep_incomplete = FALSE, tz = "Etc/GMT-1")
}
\arguments{
\item{data}{...}

\item{keep_incomplete}{...}

\item{tz}{...}
}
\description{
Calculate O3 peak-season metric from Ostluft monitoring h1 data
}
\keyword{internal}
