% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{recode_nabel_sites}
\alias{recode_nabel_sites}
\title{Recode NABEL short site names to long names}
\usage{
recode_nabel_sites(site)
}
\arguments{
\item{site}{...}
}
\description{
Recode NABEL short site names to long names
}
\keyword{internal}
