% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{recode_ostluft_meta_type}
\alias{recode_ostluft_meta_type}
\title{Recode Ostluft air quality monitoring traffic site classification}
\usage{
recode_ostluft_meta_type(type)
}
\arguments{
\item{type}{...}
}
\description{
... to be roughly in line with https://www.bafu.admin.ch/bafu/de/home/themen/luft/publikationen-studien/publikationen/immissionsmessung-von-luftfremdstoffen.html
however, the Ostluft site classes are - as categories - not entirely consistent with the current Immissionsmessempfehlung. We will need to put future effort in a reclassification.
}
\keyword{internal}
