% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{remove_duplicate_d1}
\alias{remove_duplicate_d1}
\title{Make sure that there are no duplicate measurements per site /day / unit in Ostluft dataset}
\usage{
remove_duplicate_d1(data)
}
\arguments{
\item{data}{...}
}
\description{
Function to make sure that there are no duplicate measurements per site / day / unit for data with interval = d1 in format rOstluft::format_rolf()
in case there have been eBC/EC values for both the 10 and 2.5 PM-fraction - then prefer the PM2.5 fraction.
}
\keyword{internal}
