% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{replace_ec}
\alias{replace_ec}
\title{Parameter replacement function for parameter EC / eBC}
\usage{
replace_ec(parameter, value)
}
\arguments{
\item{parameter}{...}

\item{value}{...}
}
\description{
Function to make sure that there are no duplicate measurements (used in remove_duplicate_y1())
in case there have been eBC/EC values for both the 10 and 2.5 PM-fraction - then prefer the PM2.5 fraction.
}
\keyword{internal}
