% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{replace_pm10}
\alias{replace_pm10}
\title{Parameter replacement function for parameter PM10}
\usage{
replace_pm10(parameter, value)
}
\arguments{
\item{parameter}{...}

\item{value}{...}
}
\description{
Function to make sure that there are no duplicate measurements (used in remove_duplicate_y1())
same for PM10 monitor and high volume sampler measurements (prefer high-volume-sampler data = reference method);
}
\keyword{internal}
