% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{replace_pm25}
\alias{replace_pm25}
\title{Parameter replacement function for parameter PM2.5}
\usage{
replace_pm25(parameter, value)
}
\arguments{
\item{parameter}{...}

\item{value}{...}
}
\description{
Function to make sure that there are no duplicate measurements (used in remove_duplicate_y1())
same for PM2.5 monitor and high volume sampler measurements (prefer high-volume-sampler data = reference method)
}
\keyword{internal}
