% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{restructure_monitoring_nabel_d1}
\alias{restructure_monitoring_nabel_d1}
\title{Convert NABEL d1 dataset into standard format}
\usage{
restructure_monitoring_nabel_d1(data, tz = "Etc/GMT-1")
}
\arguments{
\item{data}{...}

\item{tz}{...}
}
\description{
Convert NABEL d1 dataset into standard format
}
\keyword{internal}
