% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{restructure_monitoring_nabel_y1}
\alias{restructure_monitoring_nabel_y1}
\title{Convert NABEL y1 dataset into standard long format}
\usage{
restructure_monitoring_nabel_y1(data, keep_incomplete = FALSE)
}
\arguments{
\item{data}{...}

\item{keep_incomplete}{...}
}
\description{
Convert NABEL y1 dataset into standard long format
}
\keyword{internal}
