% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{restructure_monitoring_ostluft}
\alias{restructure_monitoring_ostluft}
\title{Convert Ostluft dataset into standard long format}
\usage{
restructure_monitoring_ostluft(
  data,
  keep_incomplete = FALSE,
  tz = "Etc/GMT-1",
  na.rm = TRUE
)
}
\arguments{
\item{data}{...}

\item{keep_incomplete}{...}

\item{tz}{...}

\item{na.rm}{...}
}
\description{
Convert Ostluft dataset into standard long format
}
\keyword{internal}
