% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{siteclass_nh3}
\alias{siteclass_nh3}
\title{Derive NH3 emission surrounding siteclass from GVE & N-Austrag within 5km}
\usage{
siteclass_nh3(gve, nfert)
}
\arguments{
\item{x}{...}
}
\description{
Derive NH3 emission surrounding siteclass from GVE & N-Austrag within 5km
}
\keyword{internal}
