% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_monitoring_met_d1}
\alias{data_monitoring_met_d1}
\title{Meteorological monitoring data within Ostluft}
\format{
## `data_monitoring_met_d1`
A tibble:
...
}
\source{
<https://www.ostluft.ch/>
}
\usage{
data_monitoring_met_d1
}
\description{
Meteorological monitoring data as daily values at sites in Ostluft, Switzerland,
for selected parameters (T, Hr, p, StrGlo, RainDur)
from cantonal (Ostluft)  air quality monitoring networks.
}
\keyword{datasets}
