% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_monitoring_ndep}
\alias{data_monitoring_ndep}
\title{Air quality monitoring data for reactive nitrogen deposition to sensitive ecosystems within the Canton of Zürich}
\format{
## `data_monitoring_ndep`
A tibble:
...
}
\source{
<https://www.ostluft.ch/>
}
\usage{
data_monitoring_ndep
}
\description{
Pre-compiled air quality monitoring data as yearly values for reactive nitrogen deposition to sensitive ecosystems
at sites in the Canton of Zürich, Switzerland,
from the cantonal (Ostluft) air quality monitoring network.
}
\keyword{datasets}
