% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_monitoring_aq}
\alias{prepare_monitoring_aq}
\title{Merge monitoring dataset with site metadata, derive pollutant and metric variable and prepare final dataset}
\usage{
prepare_monitoring_aq(data, meta)
}
\arguments{
\item{data}{...}

\item{meta}{...}
}
\description{
Merge monitoring dataset with site metadata, derive pollutant and metric variable and prepare final dataset
}
\keyword{internal}
