% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_helpers.R
\name{remove_duplicate_y1}
\alias{remove_duplicate_y1}
\title{Make sure that there are no duplicate measurements per site / year / unit in Ostluft dataset}
\usage{
remove_duplicate_y1(data)
}
\arguments{
\item{data}{...}
}
\description{
Function to make sure that there are no duplicate measurements per site / year / unit for data with interval = y1 in format rOstluft::format_rolf()
in case there have been NO2 monitor and passive sampler measurements (prefer monitor data = reference method);
same for PM10 monitor and high volume sampler measurements (prefer high-volume-sampler data = reference method);
same for PM2.5 monitor and high volume sampler measurements (prefer high-volume-sampler data = reference method)
}
\keyword{internal}
