% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_meta_ndep}
\alias{site_meta_ndep}
\title{Site-metadata for reactive nitrogen deposition to sensitive ecosystems in Ostluft}
\format{
## `site_meta_ndep`
A tibble:
...
}
\source{
<https://www.ostluft.ch/>
}
\usage{
site_meta_ndep
}
\description{
metadata at ecosystem reactive nitrogen deposition sites in the area of Ostluft, Switzerland,
from the intercantonal (Ostluft) and national (NABEL) air quality monitoring networks as well as from Swiss federal research institute WSL, in one case..
}
\keyword{datasets}
